complete

+++ admin/logic/delete-category-logic.php
@@ -6,12 +6,14 @@ if (isset($_GET['id'])) {
 
 
     // update category_id of posts that belong to this category to id of uncategorized category 
-
-
-    // delete category 
-    $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
-    $result = mysqli_query($connection, $query);
-    $_SESSION['delete-category-success'] = "Category deleted successfully";
+    $update_query = "UPDATE post SET category_id=2 WHERE category_id=$id";
+    $update_result = mysqli_query($connection, $update_query);
+    if (!mysqli_errno($connection)) {
+        // delete category 
+        $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
+        $result = mysqli_query($connection, $query);
+        $_SESSION['delete-category-success'] = "Category deleted successfully";
+    }
 }
 
 header('location: ' . ROOT_URL . 'admin/manage-categories.php');
